% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act.r
\docType{data}
\name{act}
\alias{act}
\title{Actigraphy data}
\format{A data frame

\describe{
  \item{actigraph_brand}{actigraph_brand}
  \item{subject_ID}{subject_ID}
  \item{interval_type}{interval_type}
  \item{interval_number}{interval_number}
  \item{duration}{duration}
  \item{efficiency}{efficiency}
  \item{sleep_time}{sleep_time}
  \item{bad}{bad}
  \item{datime_start}{datime_start}
  \item{datime_end}{datime_end}
  \item{comments}{comments}


}}
\source{
\url{www.sleepwake.ac.nz}
}
\usage{
act
}
\description{
Contains data from three participants.
}
\examples{
data("act")
}
