% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-actigraphs.r
\docType{data}
\name{actigraphs}
\alias{actigraphs}
\title{actigraphs data}
\format{A data frame

\describe{
  \item{actigraph}{actigraph}
  \item{analysis_name}{analysis_name}
  \item{subject_ID}{subject_ID}
  \item{interval_type}{interval_type}
  \item{interval_number}{interval_number}
  \item{start_date}{start_date}
  \item{start_time}{start_time}
  \item{end_date}{end_date}
  \item{end_time}{end_time}
  \item{duration}{duration}
  \item{efficiency}{efficiency}
  \item{sleep_time}{sleep_time}
  \item{file_name}{file_name}
  \item{bad}{bad}
  \item{datime_start}{datime_start}
  \item{datime_end}{datime_end}


}}
\source{
\url{http://www.putURL/}
}
\usage{
actigraphs
}
\description{
Contains data from different actigraphs e.g. AMI, Actiware
}
\examples{
data("actigraphs")
}

