% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_startend.r
\name{int_startend}
\alias{int_startend}
\title{Returns the starting/ending point of a report}
\usage{
int_startend(x, tz = "UTC", ...)
}
\arguments{
\item{x}{a dataframe}

\item{tz}{is the time zone}

\item{...}{Optional parameters}
}
\value{
a dataframe
}
\description{
Returns the starting/ending point of a report
}
