% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/night.shade.r
\name{local.night.shade}
\alias{local.night.shade}
\title{Generates local night time periods}
\usage{
local.night.shade(localTZ, shadow.start = "20:00:00",
  shadow.end = "06:00:00", tz = "UTC", ...)
}
\arguments{
\item{localTZ}{a dataframe containing the local time zone by participant during the days worked.}

\item{shadow.start}{starting time of the home night, 22:00:00 by defaults}

\item{shadow.end}{ending time of the home night, 08:00:00 by defaults}

\item{...}{Optional parameters}

\item{x}{a dataframe}
}
\value{
a dataframe
}
\description{
This function is used by the Darwent plot
}
\examples{
#flight <- read.csv(file = 'C:\\\\1\\\\EXAMPLE_DATA\\\\Work\\\\work.csv', sep = ",",header = TRUE, skip = 0)
#flight$datime_start <- paste( as.POSIXct( strptime( flight$StartDatime, format = "\%d/\%m/\%Y \%H:\%M"), tz = "UTC"))
#flight$datime_end <- paste( as.POSIXct( strptime( flight$EndDatime, format = "\%d/\%m/\%Y \%H:\%M"), tz = "UTC"))
#flight <- flight[,-c(3,4)]
#colnames(flight) <- c("subject_ID", "startTZ", "endTZ", "interval_type","datime_start", "datime_end")
# create a data.frame with the worked days and the tz per ID.
#localTZ <- filter(flight, interval_type == "Work") \%>\% select(subject_ID, datime_start, datime_end, startTZ)
#names(localTZ)[4] <- "TZ"
#local.night.shade(localTZ = localTZ, shadow.start = "22:00:00", shadow.end = "08:00:00")
}
