% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_period.r
\name{param_period}
\alias{param_period}
\title{Imports parameter periods}
\usage{
param_period(filepath)
}
\arguments{
\item{filepath}{the path to the file}
}
\value{
a dataframe
}
\description{
Imports parameter periods
}

