% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_point.r
\name{param_point}
\alias{param_point}
\title{Input reports based on points.}
\usage{
param_point(filepath)
}
\arguments{
\item{filepath}{the path to the file}
}
\value{
a dataframe
}
\description{
Input reports based on points.
}
\examples{
#fp2 <- 'input parameters example point.csv'
#input.param.point <- param_point (filepath = fp2)
#input.param.point$time_point_datime <- dmy_hm(input.param.point$time_point_datime,  tz = "UTC")
#input.param.point
}

