% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portion.r
\name{portion_withoverlaps}
\alias{portion_withoverlaps}
\title{Slices a data.frame containing date-time type columns including overlaping periods.}
\usage{
portion_withoverlaps(x, from, to, ...)
}
\arguments{
\item{x}{a dataframe}

\item{from}{a vector of lenght = participants}

\item{to}{a vector of lenght = participants}

\item{...}{Optional parameters}
}
\value{
a dataframe
}
\description{
Slices a data.frame containing date-time type columns including overlaping periods.
}
\details{
it does include overlaps periods: periods that started (ended) before (after) start.period (end.period)
}
\examples{
acti_data <- read_actigraph_csv(x = "C:\\\\1\\\\EXAMPLE_DATA")
start.period <- rep("2016-07-10 12:00", 3)
end.period <- rep("2016-07-20 12:00", 3)
mystudy <- portion_withoverlaps(acti_data, from = start.period,  to = end.period)
}

