% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_actigraph_csv.r
\name{read_actigraph_csv}
\alias{read_actigraph_csv}
\title{Imports scored actigraphy data from different actigraph brand output files.}
\usage{
read_actigraph_csv(x = "PATH_TO_DATA", tz = "UTC", ...)
}
\arguments{
\item{x}{the file path}

\item{tz}{the time zone}

\item{...}{Optional parameters}
}
\value{
a dataframe
}
\description{
Specify the path to the files. So far it supports CSV files only.
The imported files are merged creating an organic single file.
If the CSV files have different number of columns the function will fill missing columns with NAs.
}
