% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_filename.r
\name{read_csv_filename}
\alias{read_csv_filename}
\title{Imports a CSV file and attaches the file path as a column.}
\usage{
read_csv_filename(filename)
}
\arguments{
\item{filename}{file path}
}
\value{
a dataframe
}
\description{
Imports a CSV file and attaches the file path as a column.
}
