% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_point.r
\name{report_point}
\alias{report_point}
\title{Generates reports based on input points.}
\usage{
report_point(period, acti_data, ...)
}
\arguments{
\item{period}{a dataframe containing participants, start and end date/time interval}

\item{acti_data}{a dataframe of the form "acti_data".}

\item{...}{Optional parameters}
}
\value{
a dataframe
}
\description{
Generates reports based on input points.
}
\examples{
input.param.point <- read.csv(file = paste("C:\\\\1\\\\EXAMPLE_DATA\\\\input parameters example point.csv"), sep = ",", header = TRUE, skip = 0)
library("lubridate")
input.param.point$time_point_datime <- dmy_hm(input.param.point$time_point_datime,  tz = "UTC")
acti_data <- read_actigraph_csv(x = "C:\\\\1\\\\EXAMPLE_DATA")
rep3 <- report_point(period = input.param.point, acti_data = acti_data)
View (rep3)
}

