% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheep_counter.r
\name{sheep_counter}
\alias{sheep_counter}
\title{Sheep counter}
\usage{
sheep_counter(dat, tz = "UTC", interv = "10 mins", datebreaks = "2 hour",
  work_data)
}
\arguments{
\item{dat}{a data frame.}

\item{tz}{the time zone.}

\item{interv}{interval of time, set as 10min.}

\item{datebreaks}{is the distance between breaks in the x-axis. "2 hour" by default.}

\item{work_data}{optional parameter used to import work data.}
}
\value{
a plot.
}
\description{
Computes the empirical probability that a group or an individual are sleeping at a given time of the day.
}
\examples{
# Example 1:
dat <- read_actigraph_csv(x = "C:\\\\1\\\\EXAMPLE_DATA_6")
sheep_counter(dat)

# Example 2: using work data (randomly generated)
datime_start <- rep( seq(ISOdatetime(2012,05,12,8,00,00, tz = "UTC"),(ISOdatetime(2012,05,12,8,00,00, tz = "UTC") + lubridate::days(3)), length.out = 4), length(unique(dat$subject_ID)) )
datime_end <- rep( seq(ISOdatetime(2012,05,12,16,00,00, tz = "UTC"),(ISOdatetime(2012,05,12,16,00,00, tz = "UTC") + lubridate::days(3)), length.out = 4), length(unique(dat$subject_ID)) )
df <- data.frame(datime_start = datime_start, datime_end = datime_end, subject_ID = rep(unique(dat$subject_ID),4))
dif1 <- rnorm(nrow(df), 1,1)
dif2 <- rnorm(nrow(df), 1,1)
df$datime_start <- df$datime_start + (3600*dif1)
df$datime_end <- df$datime_end + (3600*dif2)
sheep_counter(dat, work_data = df)

# Example 3:
dat2 <- read_actigraph_csv(x = "C:\\\\1\\\\EXAMPLE_DATA_7")
sheep_counter(dat2)
}
