% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sleep_work.r
\docType{data}
\name{sleep_work}
\alias{sleep_work}
\title{sleep/work data (not real)}
\format{A data frame

\describe{
  \item{subject_ID}{subject_ID}
  \item{startTZ}{the time zone when the work period started}
  \item{endTZ}{the time zone when the work period ended}
  \item{interval_type}{interval_type}
  \item{datime_start}{datime_start}
  \item{datime_end}{datime_end}

}}
\source{
\url{http://www.putURL/}
}
\usage{
sleep_work
}
\description{
It contains sleep/work date-time intervals of 3 participants.
It includes the time zone associated with the start/end of the work periods.
}
\examples{
data("sleep_work")
}

