% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_actigraph_csv.r
\name{write_actigraph_csv}
\alias{write_actigraph_csv}
\title{Exports data as CSV.}
\usage{
write_actigraph_csv(x, path = "~", ...)
}
\arguments{
\item{x}{the file path}

\item{path}{is the path for saving the file}

\item{...}{Optional parameters}
}
\value{
a dataframe
}
\description{
Exports data as CSV.
}
