% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deagg.R
\name{deagg}
\alias{deagg}
\title{Deaggregation of reports for summary_type sequential or daily.}
\usage{
deagg(period)
}
\arguments{
\item{period}{a dataframe containing participants, start and end date/time period}
}
\value{
a dataframe
}
\description{
Deaggregation of reports for summary_type sequential or daily.
}
\examples{
library("lubridate")
library("dplyr")
p4 <- read.csv(file = paste("C:\\\\1\\\\EXAMPLE_DATA_2\\\\input parameters example period mixed.csv"), sep = ",", header = TRUE, skip = 0)
p4$summary_start_datime <- lubridate::dmy_hm(p4$summary_start_datime,  tz = "UTC")
p4$summary_end_datime <- lubridate::dmy_hm(p4$summary_end_datime,  tz = "UTC")
d <- deagg(period = p4)

}
