% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/night.shade.r
\name{home.night.shade}
\alias{home.night.shade}
\title{Generates home night time periods}
\usage{
home.night.shade(x, shadow.start = "22:00:00", shadow.end = "08:00:00",
  TZ = 0, ...)
}
\arguments{
\item{x}{a dataframe}

\item{shadow.start}{starting time of the home night, 22:00:00 by defaults}

\item{shadow.end}{ending time of the home night, 08:00:00 by defaults}

\item{TZ}{participant's home time zone}

\item{...}{Optional parameters}
}
\value{
a dataframe
}
\description{
This function is used by the Darwent plot
}
\examples{
# home.night.shade(x = flight, shadow.start = "22:00:00",
# shadow.end = "08:00:00")
}

