% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/night.shade.r
\name{local.night.shade}
\alias{local.night.shade}
\title{Generates local night time periods}
\usage{
local.night.shade(x, part_homeTZ, shadow.start = "22:00:00",
  shadow.end = "8:00:00", ...)
}
\arguments{
\item{x}{a dataframe}

\item{part_homeTZ}{a dataframe containing the participant's home time zone.}

\item{shadow.start}{starting time of the home night, 22:00:00 by defaults}

\item{shadow.end}{ending time of the home night, 08:00:00 by defaults}

\item{...}{Optional parameters}
}
\value{
a dataframe
}
\description{
This function is used by the Darwent plot
}
\examples{
flight <- read.csv(file = 'C:\\\\1\\\\EXAMPLE_DATA\\\\Work\\\\work.csv', sep = ",",header = TRUE, skip = 0)
flight$datime_start <- paste( as.POSIXct( strptime( flight$StartDatime, format = "\%d/\%m/\%Y \%H:\%M"), tz = "UTC"))
flight$datime_end <- paste( as.POSIXct( strptime( flight$EndDatime, format = "\%d/\%m/\%Y \%H:\%M"), tz = "UTC"))
flight <- flight[,-c(3,4)]
colnames(flight) <- c("subject_ID", "startTZ", "endTZ", "interval_type","datime_start", "datime_end")
library("dplyr")
flight <- dplyr::tbl_df(flight)
part_homeTZ <- data.frame (subject_ID = c("example01", "example01_AMI", "example02"), homeTZ = rep(2,3), decal = c(0,0,5))
sh <- local.night.shade(x = flight, part_homeTZ = part_homeTZ, shadow.start = "22:00:00", shadow.end = "8:00:00")
}

