% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portion.r
\name{portion}
\alias{portion}
\title{#Slices a data.frame containing date-time type columns.}
\usage{
portion(x, from, to, ...)
}
\arguments{
\item{x}{at dataframe}

\item{from}{a vector of lenght = participants}

\item{to}{a vector of lenght = participants}

\item{...}{Optional parameters}
}
\value{
a dataframe
}
\description{
#Slices a data.frame containing date-time type columns.
}
\details{
it does not include overlaps periods: periods that started (ended) before (after) start.period (end.period)
}
\examples{
acti_data <- read_actigraph_csv(x = "C:\\\\1\\\\EXAMPLE_DATA")
start.period <- rep("2016-07-10", 3)
end.period <- rep("2016-07-20", 3)
mystudy <- portion(acti_data, from = start.period,  to = end.period)

}

