% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portion.r
\name{portion_withoverlaps}
\alias{portion_withoverlaps}
\title{Slices a data.frame containing date-time type columns including overlaping periods.}
\usage{
portion_withoverlaps(x, from, to, tz = "UTC", ...)
}
\arguments{
\item{x}{a dataframe}

\item{from}{a vector of lenght = participants}

\item{to}{a vector of lenght = participants}

\item{tz}{is the time zone}

\item{...}{Optional parameters}
}
\value{
a dataframe
}
\description{
Slices a data.frame containing date-time type columns including overlaping periods.
}
\details{
it does include overlaps periods: periods that started (ended) before (after) start.period (end.period)
}
\examples{
library("lubridate")
start.period <- ymd_hms("2017-12-10 12:00:00")
end.period <- ymd_hms("2017-12-12 12:00:00")
portion_withoverlaps(act[act$subject_ID==1,], from = start.period,  to = end.period)
}
